Tools: setup-unit: ctl edit: Print file name on error
When editing the configuration in-place, it's easy to make a mistake.
If the configuration is wrong, it will be passed to the control socket,
which will reject it, keeping the old configuration.  Those manual edits
would be lost, which can make it very uncomfortable to edit in-place.

By printing the name of the temporary file, we allow the user to recover
the changes.

Cc: Liam Crilly <liam@nginx.com>
Reviewed-by: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>