Packages: added wasm-wasi-component module packaging for deb-based distros
We need to redefine CFLAGS to drop missing-prototypes as warning since
third-party code such as wasmtime fails to build from source when
building a debian package.  This happens only for packages because we
use DPKG_EXPORT_BUILDFLAGS=1 propagating build flags to the environment,
so cargo build picks it up as well.  Since we have no control over
third-party code, the easiest solution is to disable this warning.