fs: Rename nxt_fs_mkdir_parent() => nxt_fs_mkdir_dirname()
"dirname" is the usual way to refer to the directory part of a path
name.  See for example dirname(1), or the dirname builtin in several
languages.  Also, in the context of mkdir(), "parents" is used to refer
to mkdir -p, which is too similar to "parent", so it can lead to
confusion.

Tested-by: Andy Postnikov <apostnikov@gmail.com>
Tested-by: Andrew Clayton <a.clayton@nginx.com>
Reviewed-by: Andrew Clayton <a.clayton@nginx.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>