http: Move chunked buffer pos pointer while parsing
Previously, Unit didn't move the buffer pointer when parsing chunked
data because the buffer was not used after sent. It's used for upstream
response. With the new requirement to support request chunked body,
the buffer might be used for pipeline request, so it's necessary to move
the pos pointer while parsing.

Reviewed-by: Andrew Clayton <a.clayton@nginx.com>