status: Use a variable to represent the status member index
In nxt_status_get() call nxt_conf_set_member() multiple times to set the
main /status json sections.

Previously this used hard coded values, 0, 1, 2 etc, if you wanted to
change the order or insert new sections it could mean renumbering all
these.

Instead use a variable to track this index which starts at 0 and is
simply incremented in each call of nxt_conf_set_member().

Currently this is only for the main outer sections, but can be
replicated for inner sections if required.

This is a preparatory patch for adding a new "modules" section at the
top.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>