status: Show list of loaded language modules
When querying the '/status' node in the control API, display the list of
currently loaded modules.

So we now get something like

  {
      "modules": {
          "python": {
              "version": "3.12.2",
              "lib": "/opt/unit/modules/python.unit.so"
          },

          "wasm": {
              "version": "0.1",
              "lib": "/opt/unit/modules/wasm.unit.so"
          },

          "wasm-wasi-component": {
              "version": "0.1",
              "lib": "/opt/unit/modules/wasm_wasi_component.unit.so"
          }
      },

      ...
  }

This can be useful for debugging to show exactly what modules Unit has
loaded _and_ from where.

Due to the module name not actually being kept around (it could be added
but would be a much more intrusive patch) we need to map module types to
their name.

I have done this using a static inline function with a switch()
statement rather say an array, so that if the nxt_app_type_t enum is
ever updated and the switch statement here isn't, the compiler will
shout at you.

Closes: https://github.com/nginx/unit/issues/1343
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>