auto, wasm-wc: Copy the .so into build/lib/unit/modules/
Normally when the language modules are built, they are copied into the
build/lib/unit/modules/ directory (as part of the build, not install
step).

This then allows Unit to find them without being installed. This is
useful for things like the pytests.

This wasn't happening for the wasm-wasi-componment language mpdule. So
we now copy it over and give it the right name as part of the make/build
process.

Reported-by: Andrei Zeliankou <zelenkov@nginx.com>
Signed-off-by: Andrew Clayton <a.clayton@nginx.com>