Flow the language module name into nxt_app_lang_module_t
The nxt_app_lang_module_t structure contains various bits of information
as obtained from the nxt_app_module_t structure that language modules
define.

One bit of information that is in the nxt_app_module_t but not in the
nxt_app_lang_module_t is the language module name.

Having this name flowed through will be useful for displaying the loaded
language modules in the /status endpoint.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>