CONTRIBUTING.md: Update the 'Git Style Guide' section
This makes it reflect current reality.

NOTE: This removes the bit about updating the changes.xml file.

For me that has been a constant source of problems. Especially when it
hasn't been done as a separate commit (makes reverting changes harder
due to this file being constantly re-worked).

This file is also usually re-worked at release time, with the re-wording
and re-ordering of items.

In my experience it is much better to leave the updating of this file to
release time when you can use 'git shortlog -e <prev release>..' as the
source for adding entries to the changelog.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>