Resolve unused assignment in nxt_term_parse()
Both clang-analyzer and coverity flagged an issue in nxt_term_parse()
that we set 'state = st_letter' but then set it to 'state = st_space'
before using it.

While we could simply remove the first assignment and placate the
analyzers, upon further analysis it seems that there is some more
cleanup that could be done in this function.

This commit addresses the above issue, subsequent commits will continue
the cleanup.

To solve the unused assignment issue we can get rid of the

  'state == st_letter'

assignment and unconditionally execute the code that was behind the

  if (state != st_letter) {

guard. If we're not handling a space then we should have either a digit
or letter.

Also, perhaps more importantly, this if () statement would never be
false at this point as state would never == st_letter.

We may as well also remove the st_letter enum value.

The src/test/nxt_term_parse_test.c still passes

  tests: [notice] term parse test passed

NOTE: Although this function is not currently used in Unit (only by
src/test/nxt_term_parse_test.c), it is probably worth cleaning it up and
solving one of the open clang-analyzer (and coverity) issues.

Signed-off-by: Andrew Clayton <a.clayton@nginx.com>